feat: correct aid type to long

+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/controller/TaskRestController.java
@@ -13,7 +13,7 @@ public class TaskRestController {
 
     // deprecated
     @RequestMapping(value = "/task/visit/video/{aid}/record", method = RequestMethod.POST)
-    public void addVisitVideoRecord(@PathVariable Integer aid) {
+    public void addVisitVideoRecord(@PathVariable Long aid) {
 //        Long userid = 3L;  // TODO set user role only and change to real userid
 //        taskService.addVisitVideoRecord(aid, userid);
     }
