+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/VideoAidTitleDao.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface VideoAidTitleDao {
 
-    List<VideoAidTitle> queryVideoAidTitle(Integer aid);
+    List<VideoAidTitle> queryVideoAidTitle(Long aid);
 
 }
