+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/VideoLogDao.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 public interface VideoLogDao {
 
-    List<VideoLog> queryVideoLogs(Integer aid, Integer start_ts, Integer end_ts,
+    List<VideoLog> queryVideoLogs(Long aid, Integer start_ts, Integer end_ts,
                                   String attr, String oldval, String newval,
                                   Integer offset, Integer ps);
 
-    Integer queryVideoLogsCount(Integer aid, Integer start_ts, Integer end_ts,
+    Integer queryVideoLogsCount(Long aid, Integer start_ts, Integer end_ts,
                                 String attr, String oldval, String newval);
 
 }
