+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/VideoStaffExDao.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface VideoStaffExDao {
 
-    List<VideoStaffEx> queryVideoStaffsByAid(Integer aid);
+    List<VideoStaffEx> queryVideoStaffsByAid(Long aid);
 
 }
