+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/user/UserFavoriteVideoDao.java
@@ -6,16 +6,16 @@ import java.util.List;
 
 public interface UserFavoriteVideoDao {
 
-    void addUserFavoriteVideo(Integer added, Long userid, Integer aid);
+    void addUserFavoriteVideo(Integer added, Long userid, Long aid);
 
-    Integer deleteUserFavoriteVideo(Long userid, Integer aid);
+    Integer deleteUserFavoriteVideo(Long userid, Long aid);
 
-    UserFavoriteVideo queryUserFavoriteVideo(Long userid, Integer aid);
+    UserFavoriteVideo queryUserFavoriteVideo(Long userid, Long aid);
 
-    Integer queryUserFavoriteVideoCount(Integer aid);
+    Integer queryUserFavoriteVideoCount(Long aid);
 
-    List<UserFavoriteVideo> queryUserFavoriteVideoUsers(Integer aid, Integer start_ts, Integer end_ts,
+    List<UserFavoriteVideo> queryUserFavoriteVideoUsers(Long aid, Integer start_ts, Integer end_ts,
                                                         String order_by, Integer desc, Integer offset, Integer ps);
 
-    Integer queryUserFavoriteVideoUsersCount(Integer aid, Integer start_ts, Integer end_ts);
+    Integer queryUserFavoriteVideoUsersCount(Long aid, Integer start_ts, Integer end_ts);
 }
