+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/TaskService.java
@@ -2,5 +2,5 @@ package com.bunnyxt.tdd.service;
 
 public interface TaskService {
 
-    void addVisitVideoRecord(Integer aid, Long userid);
+    void addVisitVideoRecord(Long aid, Long userid);
 }
