+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/VideoAidTitleService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface VideoAidTitleService {
 
-    List<VideoAidTitle> queryVideoAidTitle(Integer aid);
+    List<VideoAidTitle> queryVideoAidTitle(Long aid);
 
 }
